requirements.txt
feather-format==0.4.0
Jinja2==2.9.6
Markdown==2.6.9
numpy==1.14.1
pandas==0.22.0
plac==0.9.6
pyarrow==0.9.0
PyYAML==3.12
